+++ src/main/java/io/github/franciscopaulinoq/zelaseveriano/infrastructure/persistence/repository/JpaCategoriaRepository.java
@@ -3,6 +3,6 @@ package io.github.franciscopaulinoq.zelaseveriano.infrastructure.persistence.rep
 import io.github.franciscopaulinoq.zelaseveriano.infrastructure.persistence.entity.CategoriaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JpaCategoriaRepository extends JpaRepository<CategoriaEntity, Long> {
+public interface JpaCategoriaRepository extends JpaRepository<CategoriaEntity, Integer> {
 
 }
